Fin du TP sans bonus

+++ resources/views/articles/show.blade.php
@@ -38,13 +38,13 @@
 
                                                             @if($a_aime == true)
                                                                 {!! Form::open(['route' => ['like.destroy', $le_like], 'method' => 'delete']) !!}
-                                                                {{ Form::hidden('article_id', $article->id) }}
-                                                                {!! Form::submit('Je n\'aime plus', ['class' => 'btn btn-info pull-right col-md-6']) !!}
+                                                                    {{ Form::hidden('article_id', $article->id) }}
+                                                                    {!! Form::submit('Je n\'aime plus', ['class' => 'btn btn-info pull-right col-md-6']) !!}
                                                                 {!! Form::close() !!}
                                                             @else
                                                                 {!! Form::open(['route' => 'like.store', 'method' => 'post', 'class'=>'register-form']) !!}
-                                                                {{ Form::hidden('article_id', $article->id) }}
-                                                                {!! Form::submit('J\'aime', ['class' => 'btn btn-info pull-right col-md-6']) !!}
+                                                                    {{ Form::hidden('article_id', $article->id) }}
+                                                                    {!! Form::submit('J\'aime', ['class' => 'btn btn-info pull-right col-md-6']) !!}
                                                                 {!! Form::close() !!}
                                                             @endif
 
@@ -54,6 +54,21 @@
                                                         @endif
                                                     </div>
                                                 </div>
+                                                <!-- Supression et edition de l'article (Si co et apartient) -->
+                                                <ul class="boutons">
+                                                    @if(Auth::check() && (Auth::user()->id == $article->user_id || Auth::user()->isAdmin == true))
+                                                        <li>
+                                                            {!! Form::open(['route' => ['article.edit', $id], 'method' => 'get']) !!}
+                                                            {!! Form::submit('Editer', ['class' => 'btn btn-info ']) !!}
+                                                            {!! Form::close() !!}
+                                                        </li>
+                                                        <li>
+                                                            {!! Form::open(['route' => ['article.destroy', $id], 'method' => 'delete']) !!}
+                                                            {!! Form::submit('Supprimer cette article', ['class' => 'btn btn-danger']) !!}
+                                                            {!! Form::close() !!}
+                                                        </li>
+                                                    @endif
+                                                </ul>
                                             </div>
                                             <div class="col-md-6 col-sm-12">
                                                 <img class="img_article" src="../images/articles/{{$article->image_path}}" alt="">
@@ -66,19 +81,19 @@
                                                 <p>{{$article->content}}</p>
                                                 <h3>Auteur : {{$article->user->name}}</h3>
                                                 <div class="row">
-                                                    <div class="col-md-6 col-sm-12">
+                                                    <div class="col-md-3 col-sm-6">
                                                         <h4>Partage :</h4>
                                                         @include('components.share', [
                                                         'url' => request()->fullUrl(),
                                                        ])
                                                     </div>
-                                                    <div class="col-md-6 col-sm-12">
+                                                    <div class="col-md-3 col-sm-6 zone_like">
                                                         @forelse($article->likes as $like)
                                                             @if($loop->last)
-                                                                {{$loop->count}}
+                                                                <div class="col-md-6 likes">{{$loop->count}} J'aimes</div>
                                                             @endif
                                                         @empty
-                                                            Pas de J'aime
+                                                            <div class="col-md-6 likes">Aucun J'aimes</div>
                                                         @endforelse
 
                                                         @if(Auth::check())
@@ -102,24 +117,26 @@
                                                         @endif
                                                     </div>
                                                 </div>
-
+                                                <!-- Supression et edition de l'article (Si co et apartient) -->
+                                                <ul class="boutons">
+                                                    @if(Auth::check() && (Auth::user()->id == $article->user_id || Auth::user()->isAdmin == true))
+                                                        <li>
+                                                            {!! Form::open(['route' => ['article.edit', $id], 'method' => 'get']) !!}
+                                                            {!! Form::submit('Editer', ['class' => 'btn btn-info ']) !!}
+                                                            {!! Form::close() !!}
+                                                        </li>
+                                                        <li>
+                                                            {!! Form::open(['route' => ['article.destroy', $id], 'method' => 'delete']) !!}
+                                                            {!! Form::submit('Supprimer cette article', ['class' => 'btn btn-danger']) !!}
+                                                            {!! Form::close() !!}
+                                                        </li>
+                                                    @endif
+                                                </ul>
                                             </div>
                                         </div>
                                     @endif
 
                                 </div><br>
-
-
-
-                        <!-- Supression de l'article (Si co et apartient) -->
-
-                        @if(Auth::check() && (Auth::user()->id == $article->user_id || Auth::user()->isAdmin == true))
-                            {!! Form::open(['route' => ['article.destroy', $id], 'method' => 'delete']) !!}
-                                {!! Form::submit('Supprimer cette article', ['class' => 'btn btn-danger pull-right delete_button']) !!}
-                            {!! Form::close() !!}
-                        @endif
-
-
 
 
                     <!-- Création commentaire de l'article (Si co) -->
@@ -146,26 +163,50 @@
                     <!--Liste de commentaire -->
 
                         @forelse($article->comments as $comment)
-                            <div class="row">
-                                <div class="col-xs-6">
-                                    <h4>{{$comment->user->name}}</h4>
-                                    <p>{{$comment->comment}}</p><br>
-                                </div>
-                                <div class="col-xs-3">
-                                    <img class="img_comment" src="../images/comments/{{$comment->comment_img}}" alt="">
-                                </div>
-                                <div class="col-xs-3">
-                                    @if(Auth::check() && $comment->user_id == Auth::user()->id)
-                                        {!! Form::open(['route' => ['commentaire.edit', $comment->id], 'method' => 'get']) !!}
-                                            {!! Form::submit('Editer', ['class' => 'btn btn-info pull-right']) !!}
-                                        {!! Form::close() !!}
-                                        {!! Form::open(['route' => ['commentaire.destroy', $comment->id], 'method' => 'delete']) !!}
-                                            {!! Form::submit('Supprimer', ['class' => 'btn btn-danger pull-right']) !!}
-                                        {!! Form::close() !!}
-                                    @endif
-                                </div>
+                            <div class="row un_commentaire">
+                                @if($comment->comment_img != null)
+                                    <div class="col-xs-6">
+                                        <h4>{{$comment->user->name}}</h4>
+                                        <p>{{$comment->comment}}</p><br>
+                                    </div>
+                                    <div class="col-xs-3">
+                                        <img class="img_comment" src="../images/comments/{{$comment->comment_img}}" alt="">
+                                    </div>
+                                    <ul class="col-xs-3 boutons">
+                                        @if(Auth::check() && $comment->user_id == Auth::user()->id)
+                                            <li>
+                                                {!! Form::open(['route' => ['commentaire.edit', $comment->id], 'method' => 'get']) !!}
+                                                {!! Form::submit('Editer', ['class' => 'btn btn-info pull-right']) !!}
+                                                {!! Form::close() !!}
+                                            </li>
+                                            <li>
+                                                {!! Form::open(['route' => ['commentaire.destroy', $comment->id], 'method' => 'delete']) !!}
+                                                {!! Form::submit('Supprimer', ['class' => 'btn btn-danger pull-right']) !!}
+                                                {!! Form::close() !!}
+                                            </li>
+                                        @endif
+                                    </ul>
+                                @else
+                                    <div class="col-xs-9">
+                                        <h4>{{$comment->user->name}}</h4>
+                                        <p>{{$comment->comment}}</p><br>
+                                    </div>
+                                    <ul class="col-xs-3 boutons">
+                                        @if(Auth::check() && $comment->user_id == Auth::user()->id)
+                                            <li>
+                                                {!! Form::open(['route' => ['commentaire.edit', $comment->id], 'method' => 'get']) !!}
+                                                    {!! Form::submit('Editer', ['class' => 'btn btn-info pull-right']) !!}
+                                                {!! Form::close() !!}
+                                            </li>
+                                            <li>
+                                                {!! Form::open(['route' => ['commentaire.destroy', $comment->id], 'method' => 'delete']) !!}
+                                                    {!! Form::submit('Supprimer', ['class' => 'btn btn-danger pull-right']) !!}
+                                                {!! Form::close() !!}
+                                            </li>
+                                        @endif
+                                    </ul>
+                                @endif
                             </div>
-
 
                         @empty
                             <p>Aucun commentaire</p>
